Reorganise to prevent circular dependencies
