+++ assets/js/main.js
@@ -1,2 +1,2 @@
-!function(){"use strict";const e={buildElement:(e,n)=>{const t=document.createElement(e);return t.classList.add(n),t},getPercentage:e=>Math.floor(100*e)};function n(n,t,a){this.currencyName=t,this.unitPrice=n/1e8,this.unitPercentage=e.getPercentage(this.unitPrice),this.smallUnitName=a,this.smallestUnitPrice=this.currencyName===this.smallUnitName?this.unitPrice:100*n/1e8,this.smallestUnitPercentage=e.getPercentage(this.smallestUnitPrice),this.smallestUnitKilled=this.currencyName!=this.smallUnitName&&this.smallestUnitPercentage>100}function t(e,t,a){switch(e){case"ARS":{const e=new n(t,"Argentine Peso","Argentine Peso");Object.assign(a,e);break}case"AUD":{const e=new n(t,"Australian Dollar","Australian Cent");Object.assign(a,e);break}case"BRL":{const e=new n(t,"Brazilian Real","Brazilian Centavo");Object.assign(a,e);break}case"CAD":{const e=new n(t,"Canadian Dollar","Canadian Cent");Object.assign(a,e);break}case"CHF":{const e=new n(t,"Swiss Franc","Swiss Rappen");Object.assign(a,e);break}case"CLP":{const e=new n(t,"Chilean Peso","Chilean Peso");Object.assign(a,e);break}case"CNY":{const e=new n(t,"Chinese Yuan","Chinese Fen");Object.assign(a,e);break}case"CZK":{const e=new n(t,"Czech Koruna","Czech Haléř");Object.assign(a,e);break}case"DKK":{const e=new n(t,"Danish Krone","Danish Øre");Object.assign(a,e);break}case"EUR":{const e=new n(t,"Euro","Euro Cent");Object.assign(a,e);break}case"GBP":{const e=new n(t,"Pound Sterling","British Penny");Object.assign(a,e);break}case"HKD":{const e=new n(t,"Hong Kong Dollar","Hong Kong Cent");Object.assign(a,e);break}case"HRK":{const e=new n(t,"Croatian Kuna","Croatian Lipa");Object.assign(a,e);break}case"HUF":{const e=new n(t,"Hungarian Forint","Hungarian Forint");Object.assign(a,e);break}case"INR":{const e=new n(t,"Indian Rupee","Indian Paisa");Object.assign(a,e);break}case"ISK":{const e=new n(t,"Icelandic Króna","Icelandic Króna");Object.assign(a,e);break}case"JPY":{const e=new n(t,"Japanese Yen","Japanese Yen");Object.assign(a,e);break}case"KRW":{const e=new n(t,"South Korean Won","South Korean Won");Object.assign(a,e);break}case"NZD":{const e=new n(t,"New Zealand Dollar","New Zealand Cent");Object.assign(a,e);break}case"PLN":{const e=new n(t,"Polish Złoty","Polish Grosz");Object.assign(a,e);break}case"RON":{const e=new n(t,"Romanian Leu","Romanian Ban");Object.assign(a,e);break}case"RUB":{const e=new n(t,"Russian Ruble","Russian Kopek");Object.assign(a,e);break}case"SEK":{const e=new n(t,"Swedish Krona","Swedish Öre");Object.assign(a,e);break}case"SGD":{const e=new n(t,"Singapore Dollar","Singapore Cent");Object.assign(a,e);break}case"THB":{const e=new n(t,"Thai Baht","Thai Satang");Object.assign(a,e);break}case"TRY":{const e=new n(t,"Turkish Lira","Turkish Kuruş");Object.assign(a,e);break}case"TWD":{const e=new n(t,"New Taiwan Dollar","Taiwanese Cent");Object.assign(a,e);break}case"USD":{const e=new n(t,"United States Dollar","United States Cent");Object.assign(a,e);break}}}const a=document.getElementById("js-app");let s;!async function(){try{s=await fetch("https://blockchain.info/ticker"),s=await s.json()}catch(e){console.log("No data")}finally{!function(){function n(){for(const e in s){let n=s[e];const a=n.symbol,c=n.sell;["15m","buy","last"].forEach((e=>{delete n[e]})),t(a,c,n)}}function c(){s=Object.entries(s),s.sort(((e,n)=>e[1].smallestUnitPercentage-n[1].smallestUnitPercentage)),s=Object.fromEntries(s)}n(),c(),function(){const n=e.buildElement("ul","o-currencygrid");for(const e in s){let n=s[e];t(n.currencyName,n.smallUnitName,n.unitPercentage,n.smallestUnitPercentage,n.smallestUnitKilled)}function t(t,a,s,c,i){const r=e.buildElement("li","c-currencyitem"),o=e.buildElement("div","c-currencyitem__main"),l=e.buildElement("span","c-currencyitem__text");let b,u;o.append(l);const g=n=>{const t=n>100?100:n,a=e.buildElement("span","c-currencyitem__percentage-bar");return a.style.width=t+"%",a};c>100&&r.classList.add("c-currencyitem--complete"),i?(r.classList.add("c-currencyitem--smallest-complete"),b=`${t} ${s}%`,u=`${a} ${c}%`,l.innerText=u,r.addEventListener("mouseenter",(()=>{l.innerText=b})),r.addEventListener("mouseleave",(()=>{l.innerText=u})),r.append(o),o.append(g(c))):(b=a<1?`${c} <1%`:`${a} ${c}%`,l.innerText=b,r.append(o),o.append(g(c))),n.append(r)}a.append(n)}()}()}}(),function(){const n="npub1y73ajm09j3wra8jj9e3h8rkj3xculh520m9plgc6j57rkg7g0yyqt5p5dt",t=`${n.substring(0,15)}...${n.slice(-15)}`,a=e.buildElement("span","c-nostr-copy__message");a.innerHTML="Copied";const s=document.getElementById("js-nostr-key"),c=document.getElementById("js-nostr-copy"),i=document.querySelector(".js-copy-icon");s.innerText=t,c.addEventListener("click",(()=>{navigator.clipboard.writeText(n).then((function(){console.log("copied"),c.append(a),i.remove(),console.log("copied"),setTimeout((()=>{c.append(i),a.remove()}),2e3)})).catch((function(e){console.error("Failed to copy text to clipboard: ",e)}))}))}()}();
+!function(){"use strict";const e={buildElement:(e,n)=>{const t=document.createElement(e);return t.classList.add(n),t},getPercentage:e=>Math.floor(100*e)};function n(n,t,a){this.currencyName=t,this.unitPrice=n/1e8,this.unitPercentage=e.getPercentage(this.unitPrice),this.smallUnitName=a,this.smallestUnitPrice=this.currencyName===this.smallUnitName?this.unitPrice:100*n/1e8,this.smallestUnitPercentage=e.getPercentage(this.smallestUnitPrice),this.smallestUnitKilled=this.currencyName!=this.smallUnitName&&this.smallestUnitPercentage>100}function t(e,t,a){switch(e){case"ARS":{const e=new n(t,"Argentine Peso","Argentine Peso");Object.assign(a,e);break}case"AUD":{const e=new n(t,"Australian Dollar","Australian Cent");Object.assign(a,e);break}case"BRL":{const e=new n(t,"Brazilian Real","Brazilian Centavo");Object.assign(a,e);break}case"CAD":{const e=new n(t,"Canadian Dollar","Canadian Cent");Object.assign(a,e);break}case"CHF":{const e=new n(t,"Swiss Franc","Swiss Rappen");Object.assign(a,e);break}case"CLP":{const e=new n(t,"Chilean Peso","Chilean Peso");Object.assign(a,e);break}case"CNY":{const e=new n(t,"Chinese Yuan","Chinese Fen");Object.assign(a,e);break}case"CZK":{const e=new n(t,"Czech Koruna","Czech Haléř");Object.assign(a,e);break}case"DKK":{const e=new n(t,"Danish Krone","Danish Øre");Object.assign(a,e);break}case"EUR":{const e=new n(t,"Euro","Euro Cent");Object.assign(a,e);break}case"GBP":{const e=new n(t,"Pound Sterling","British Penny");Object.assign(a,e);break}case"HKD":{const e=new n(t,"Hong Kong Dollar","Hong Kong Cent");Object.assign(a,e);break}case"HRK":{const e=new n(t,"Croatian Kuna","Croatian Lipa");Object.assign(a,e);break}case"HUF":{const e=new n(t,"Hungarian Forint","Hungarian Forint");Object.assign(a,e);break}case"INR":{const e=new n(t,"Indian Rupee","Indian Paisa");Object.assign(a,e);break}case"ISK":{const e=new n(t,"Icelandic Króna","Icelandic Króna");Object.assign(a,e);break}case"JPY":{const e=new n(t,"Japanese Yen","Japanese Yen");Object.assign(a,e);break}case"KRW":{const e=new n(t,"South Korean Won","South Korean Won");Object.assign(a,e);break}case"NZD":{const e=new n(t,"New Zealand Dollar","New Zealand Cent");Object.assign(a,e);break}case"PLN":{const e=new n(t,"Polish Złoty","Polish Grosz");Object.assign(a,e);break}case"RON":{const e=new n(t,"Romanian Leu","Romanian Ban");Object.assign(a,e);break}case"RUB":{const e=new n(t,"Russian Ruble","Russian Kopek");Object.assign(a,e);break}case"SEK":{const e=new n(t,"Swedish Krona","Swedish Öre");Object.assign(a,e);break}case"SGD":{const e=new n(t,"Singapore Dollar","Singapore Cent");Object.assign(a,e);break}case"THB":{const e=new n(t,"Thai Baht","Thai Satang");Object.assign(a,e);break}case"TRY":{const e=new n(t,"Turkish Lira","Turkish Kuruş");Object.assign(a,e);break}case"TWD":{const e=new n(t,"New Taiwan Dollar","Taiwanese Cent");Object.assign(a,e);break}case"USD":{const e=new n(t,"United States Dollar","United States Cent");Object.assign(a,e);break}}}!function(){const n=document.getElementById("js-app");let a;!async function(){try{a=await fetch("https://blockchain.info/ticker"),a=await a.json()}catch(e){console.log("No data")}finally{!function(){function s(){for(const e in a){let n=a[e];const s=n.symbol,c=n.sell;["15m","buy","last"].forEach((e=>{delete n[e]})),t(s,c,n)}}function c(){a=Object.entries(a),a.sort(((e,n)=>e[1].smallestUnitPercentage-n[1].smallestUnitPercentage)),a=Object.fromEntries(a)}function i(){const t=e.buildElement("ul","o-currencygrid");for(const e in a){let n=a[e];s(n.currencyName,n.smallUnitName,n.unitPercentage,n.smallestUnitPercentage,n.smallestUnitKilled)}function s(n,a,s,c,i){const r=e.buildElement("li","c-currencyitem"),o=e.buildElement("div","c-currencyitem__main"),l=e.buildElement("span","c-currencyitem__text");let b,u;o.append(l);const g=n=>{const t=n>100?100:n,a=e.buildElement("span","c-currencyitem__percentage-bar");return a.style.width=t+"%",a};c>100&&r.classList.add("c-currencyitem--complete"),i?(r.classList.add("c-currencyitem--smallest-complete"),b=`${n} ${s}%`,u=`${a} ${c}%`,l.innerText=u,r.addEventListener("mouseenter",(()=>{l.innerText=b})),r.addEventListener("mouseleave",(()=>{l.innerText=u})),r.append(o),o.append(g(c))):(b=a<1?`${c} <1%`:`${a} ${c}%`,l.innerText=b,r.append(o),o.append(g(c))),t.append(r)}n.append(t)}s(),c(),i()}()}}()}(),function(){const n="npub1y73ajm09j3wra8jj9e3h8rkj3xculh520m9plgc6j57rkg7g0yyqt5p5dt",t=`${n.substring(0,15)}...${n.slice(-15)}`,a=e.buildElement("span","c-nostr-copy__message");a.innerHTML="Copied";const s=document.getElementById("js-nostr-key"),c=document.getElementById("js-nostr-copy"),i=document.querySelector(".js-copy-icon");s.innerText=t,c.addEventListener("click",(()=>{navigator.clipboard.writeText(n).then((function(){console.log("copied"),c.append(a),i.remove(),console.log("copied"),setTimeout((()=>{c.append(i),a.remove()}),2e3)})).catch((function(e){console.error("Failed to copy text to clipboard: ",e)}))}))}()}();
 //# sourceMappingURL=main.js.map
