+++ src/js/data-fetch-add.js
@@ -1,0 +1,111 @@
+import addCurrencyDetails from "./add-currency-details";
+import utilities from "./utilities";
+
+function fetchModifySortAdd(){
+
+    // DOM elements
+    const container = document.getElementById('js-app');
+    // Data
+    const tickerFetchUrl = "https://blockchain.info/ticker";
+
+    let fetchedData;
+
+    async function fetchData(){
+        try {
+            fetchedData = await fetch(tickerFetchUrl);
+            fetchedData = await fetchedData.json();
+        } catch (error) {
+            console.log('No data');
+        } finally {
+            dataFetched();
+        }
+    }
+
+    function dataFetched(){
+
+        // Modify data
+        function modifyData(){
+            for (const currencyItem in fetchedData) {
+                let details = fetchedData[currencyItem];
+                const symbol = details['symbol'];
+                const price = details['sell'];
+                const deleteDeets = ['15m', 'buy', 'last'];
+                deleteDeets.forEach((item) => {
+                    delete details[item];
+                })
+                addCurrencyDetails(symbol, price, details);
+            }
+        };
+        modifyData();
+
+        // Sort data
+        function sortData(){
+            fetchedData = Object.entries(fetchedData);
+            fetchedData.sort((a, b) => a[1].smallestUnitPercentage - b[1].smallestUnitPercentage);
+            fetchedData = Object.fromEntries(fetchedData);
+        };
+        sortData();
+
+        // Add data
+        function addToDOM(){
+            const currencyGrid = utilities.buildElement('ul', 'o-currencygrid');
+        
+            for (const currencyItem in fetchedData) {
+                let details = fetchedData[currencyItem];
+                addCurrency(details['currencyName'], details['smallUnitName'], details['unitPercentage'], details['smallestUnitPercentage'], details['smallestUnitKilled']);
+            }
+        
+            function addCurrency(unitName, smallUnitName, unitPercentage, smallestUnitPercentage,  smallestUnitKilled){
+                const item = utilities.buildElement('li', 'c-currencyitem');
+                const itemMain = utilities.buildElement('div', 'c-currencyitem__main');
+                const itemText = utilities.buildElement('span', 'c-currencyitem__text');
+                let itemTextString;
+                let itemTextStringsmall;
+                itemMain.append(itemText);
+                const percentageBar = (percentage) => {
+                    const width = percentage > 100 ? 100 : percentage;
+                    const element = utilities.buildElement('span', 'c-currencyitem__percentage-bar');
+                    element.style.width = width + '%';
+                    return element;
+                };
+                if (smallestUnitPercentage > 100) {
+                    item.classList.add('c-currencyitem--complete');
+                }
+                if (smallestUnitKilled){
+                    item.classList.add('c-currencyitem--smallest-complete');
+                    // main title
+                    itemTextString = `${unitName} ${unitPercentage}%`;
+                    itemTextStringsmall = `${smallUnitName} ${smallestUnitPercentage}%`;
+                    itemText.innerText = itemTextStringsmall;
+                    item.addEventListener('mouseenter',() => {
+                        itemText.innerText = itemTextString;
+                    });
+                    item.addEventListener('mouseleave',() => {
+                        itemText.innerText = itemTextStringsmall;
+                    });
+                    item.append(itemMain);
+                    itemMain.append(percentageBar(smallestUnitPercentage));
+                } else {
+                    // main title
+                    itemTextString = smallUnitName < 1 ? `${smallestUnitPercentage} <1%` : `${smallUnitName} ${smallestUnitPercentage}%`;
+                    itemText.innerText = itemTextString;
+                    item.append(itemMain);
+                    itemMain.append(percentageBar(smallestUnitPercentage));
+                }
+        
+                currencyGrid.append(item);
+            }
+        
+            container.append(currencyGrid);
+        
+        }
+        addToDOM();
+
+    }
+
+
+    fetchData();
+
+}
+
+export default fetchModifySortAdd;
